Convert variables involving subtraction of small values to 'c_double'; fix snow_storage to include interception; make calculation of FAO56 Ke the same as for SWB 1.0.

+++ src/actual_et__fao56.F90
@@ -57,7 +57,7 @@ contains
 
 
     real (kind=c_double), intent(inout)            :: actual_et
-    real (kind=c_float), intent(inout)             :: adjusted_depletion_fraction_p
+    real (kind=c_double), intent(inout)            :: adjusted_depletion_fraction_p
     real (kind=c_float), intent(in)                :: depletion_fraction_p
     real (kind=c_double), intent(in)               :: soil_storage
     real (kind=c_float), intent(in)                :: soil_storage_max
